move checking and reservations to tables.py, remove unused files
